Refactor: auth로 이동
기존의 로그인, 회원가입 로직들 다 auth 아래로 이동
